migrate checkRuleJoinCondition function from service layer to domain layer + Add new Domain Test & Integration Test
