+++ src/main/java/com/peoplein/moiming/domain/Moim.java
@@ -1,6 +1,9 @@
 package com.peoplein.moiming.domain;
 
 import com.peoplein.moiming.domain.embeddable.Area;
+import com.peoplein.moiming.domain.enums.MemberGender;
+import com.peoplein.moiming.domain.enums.MoimMemberState;
+import com.peoplein.moiming.domain.enums.MoimRoleType;
 import com.peoplein.moiming.domain.rules.MoimRule;
 import com.peoplein.moiming.domain.rules.RuleJoin;
 import com.peoplein.moiming.domain.rules.RulePersist;
@@ -17,7 +20,7 @@ import java.util.List;
 @Getter
 @Table(name = "moim")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Moim {
+public class Moim extends BaseEntity{
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
@@ -38,10 +41,8 @@ public class Moim {
     @Embedded
     private Area moimArea;
 
-    private LocalDateTime createdAt;
     private String createdUid;
 
-    private LocalDateTime updatedAt;
     private String updatedUid;
 
     @OneToMany(mappedBy = "moim", cascade = CascadeType.ALL)
@@ -71,7 +72,6 @@ public class Moim {
         /*
          기본값
          */
-        this.createdAt = LocalDateTime.now();
         this.hasRulePersist = false;
         this.curMemberCount = 0;
     }
@@ -149,7 +149,60 @@ public class Moim {
         this.moimInfo = moimInfo;
     }
 
-    public void setUpdatedAt(LocalDateTime updatedAt) {
-        this.updatedAt = updatedAt;
+
+    public MoimMemberState checkRuleJoinCondition(MemberInfo memberInfo, List<MemberMoimLinker> memberMoimLinkers) {
+
+        RuleJoin ruleJoin = this.getRuleJoin();
+
+        // 1. 생년월일 판별
+        if (ruleJoin.getBirthMax() != 0 && ruleJoin.getBirthMin() != 0) { // 판별조건이 있다면
+            if (memberInfo.getMemberBirth().getYear() < ruleJoin.getBirthMin()
+                    || memberInfo.getMemberBirth().getYear() > ruleJoin.getBirthMax()) {
+                return MoimMemberState.WAIT_BY_AGE;
+            }
+        }
+
+        // 2. 성별 판별
+        if (ruleJoin.getGender() != MemberGender.N) { // 판별조건이 있다면
+            if (ruleJoin.getGender() != memberInfo.getMemberGender()) {
+                return MoimMemberState.WAIT_BY_GENDER;
+            }
+        }
+
+        if (!ruleJoin.isDupLeaderAvailable() || !ruleJoin.isDupManagerAvailable() || ruleJoin.getMoimMaxCount() > 0) { // 겸직 조건이 하나라도 있거나 최대 모임 갯수 조건이 있음
+            boolean isMemberAnyLeader = false;
+            boolean isMemberAnyManager = false;
+            int cntInactiveMoim = 0;
+
+            for (MemberMoimLinker memberMoimLinker : memberMoimLinkers) {
+                if (memberMoimLinker.getMoimRoleType().equals(MoimRoleType.LEADER)) {
+                    isMemberAnyLeader = true;
+                }
+                if (memberMoimLinker.getMoimRoleType().equals(MoimRoleType.MANAGER)) {
+                    isMemberAnyManager = true;
+                }
+                if (memberMoimLinker.getMemberState() != MoimMemberState.ACTIVE) {
+                    cntInactiveMoim++;
+                }
+            }
+
+            // 3. 겸직 여부 판별
+            if (!ruleJoin.isDupLeaderAvailable()) { // 모임장 겸직 금지인데
+                if (isMemberAnyLeader) return MoimMemberState.WAIT_BY_DUP;
+            }
+
+            if (!ruleJoin.isDupManagerAvailable()) { // 운영진 겸직 금지인데
+                if (isMemberAnyManager) return MoimMemberState.WAIT_BY_DUP;
+            }
+
+            // 4. 가입 모임 수 제한
+            if (ruleJoin.getMoimMaxCount() <= memberMoimLinkers.size() - cntInactiveMoim) {
+                return MoimMemberState.WAIT_BY_MOIM_CNT;
+            }
+        }
+
+        // 모임 가입 조건 충족시 그냥 가입된다
+        return MoimMemberState.ACTIVE;
     }
+
 }
