+++ src/test/java/com/peoplein/moiming/service/MoimMemberServiceTest.java
@@ -191,182 +191,8 @@ public class MoimMemberServiceTest {
 //        assertEquals(2, curMoim.getMemberMoimLinkers().size());
     }
 
-    @Test
-    @DisplayName("성공 @ requestJoin() - 모든 RuleJoin 조건에 충족하는 모임에 가입 요청")
-    void 모든_조건_충족_가입요청() {
-
-        // given (Rule Join 이 존재하도록 설정)
-        when(curMoimRuleJoin.getBirthMax()).thenReturn(2000);
-        when(curMoimRuleJoin.getBirthMin()).thenReturn(1990);
-        when(curMoimRuleJoin.getGender()).thenReturn(MemberGender.M);
-        when(curMoimRuleJoin.getMoimMaxCount()).thenReturn(2);
-        when(curMoimRuleJoin.isDupLeaderAvailable()).thenReturn(false);
-        when(curMoimRuleJoin.isDupManagerAvailable()).thenReturn(true);
-
-        // given - SAMPLE MOIM LINKERS
-        List<MemberMoimLinker> sampleList = new ArrayList<>();
-        MemberMoimLinker linker1 = mock(MemberMoimLinker.class);
-        MemberMoimLinker linker2 = mock(MemberMoimLinker.class);
-        when(linker1.getMoimRoleType()).thenReturn(MoimRoleType.NORMAL);
-        when(linker1.getMemberState()).thenReturn(MoimMemberState.ACTIVE);
-        when(linker2.getMoimRoleType()).thenReturn(MoimRoleType.MANAGER); // 매니저 겸직 허용 충족
-        when(linker2.getMemberState()).thenReturn(MoimMemberState.IBF);
-        sampleList.add(linker1);
-        sampleList.add(linker2);
-
-        // given - 사용 repository 함수 지정
-        when(moimRepository.findWithRulesById(curMoim.getId())).thenReturn(curMoim);
-        when(memberMoimLinkerRepository.findByMemberId(requestMember.getId())).thenReturn(sampleList);
-        when(memberRepository.findMemberAndMemberInfoById(requestMember.getId())).thenReturn(requestMember);
-
-        // given (request 정보)
-        MoimJoinRequestDto requestDto = new MoimJoinRequestDto(curMoim.getId());
-
-        // when
-        MyMoimLinkerDto myMoimLinkerDto = moimMemberService.requestJoin(requestDto, requestMember);
-
-        // then
-        assertEquals(MoimRoleType.NORMAL, myMoimLinkerDto.getMoimRoleType());
-        assertEquals(MoimMemberState.ACTIVE, myMoimLinkerDto.getMemberState());
-        assertNotNull(myMoimLinkerDto.getCreatedAt());
-    }
-
-    @Test
-    @DisplayName("실패 @ requestJoin() - 나이 초과로 가입 요청 대기")
-    void 나이초과_가입요청() {
-
-        // given (Rule Join 이 존재하도록 설정) 나이가 많아서일수도 적어서일수도
-//        when(curMoimRuleJoin.getBirthMax()).thenReturn(2000);
-//        when(curMoimRuleJoin.getBirthMin()).thenReturn(1995);
-        when(curMoimRuleJoin.getBirthMax()).thenReturn(1990);
-        when(curMoimRuleJoin.getBirthMin()).thenReturn(1985);
-
-        // given - 사용 repository 함수 지정
-        lenient().when(moimRepository.findWithRulesById(curMoim.getId())).thenReturn(curMoim);
-        lenient().when(memberRepository.findMemberAndMemberInfoById(requestMember.getId())).thenReturn(requestMember);
-
-        // given (request 정보)
-        MoimJoinRequestDto requestDto = new MoimJoinRequestDto(curMoim.getId());
-
-        // when
-        MyMoimLinkerDto myMoimLinkerDto = moimMemberService.requestJoin(requestDto, requestMember);
-
-        // then
-        assertEquals(MoimRoleType.NORMAL, myMoimLinkerDto.getMoimRoleType());
-        assertEquals(MoimMemberState.WAIT_BY_AGE, myMoimLinkerDto.getMemberState());
-        assertNotNull(myMoimLinkerDto.getCreatedAt());
-    }
 
 
-    @Test
-    @DisplayName("실패 @ requestJoin() - 요구 성별 불일치로 가입 요청 대기")
-    void 성별불일치_가입요청() {
-
-        // given (Rule Join 이 존재하도록 설정)
-        when(curMoimRuleJoin.getGender()).thenReturn(MemberGender.F);
-
-        // given - 사용 repository 함수 지정
-        when(moimRepository.findWithRulesById(curMoim.getId())).thenReturn(curMoim);
-        when(memberRepository.findMemberAndMemberInfoById(requestMember.getId())).thenReturn(requestMember);
-
-        // given (request 정보)
-        MoimJoinRequestDto requestDto = new MoimJoinRequestDto(curMoim.getId());
-
-        // when
-        MyMoimLinkerDto myMoimLinkerDto = moimMemberService.requestJoin(requestDto, requestMember);
-
-        // then
-        assertEquals(MoimRoleType.NORMAL, myMoimLinkerDto.getMoimRoleType());
-        assertEquals(MoimMemberState.WAIT_BY_GENDER, myMoimLinkerDto.getMemberState());
-        assertNotNull(myMoimLinkerDto.getCreatedAt());
-    }
-
-    @Test
-    @DisplayName("실패 @ requestJoin() - 가입 모임수 초과로 가입 요청 대기")
-    void 가입_모임수_초과_가입요청() {
-
-        // given (Rule Join 이 존재하도록 설정)
-        when(curMoimRuleJoin.getMoimMaxCount()).thenReturn(2);
-
-        // given - SAMPLE MOIM LINKERS
-        List<MemberMoimLinker> sampleList = new ArrayList<>();
-        MemberMoimLinker linker1 = mock(MemberMoimLinker.class);
-        MemberMoimLinker linker2 = mock(MemberMoimLinker.class);
-        MemberMoimLinker linker3 = mock(MemberMoimLinker.class);
-        when(linker1.getMoimRoleType()).thenReturn(MoimRoleType.NORMAL);
-        when(linker1.getMemberState()).thenReturn(MoimMemberState.WAIT);
-        when(linker2.getMoimRoleType()).thenReturn(MoimRoleType.NORMAL);
-        when(linker2.getMemberState()).thenReturn(MoimMemberState.ACTIVE);
-        when(linker3.getMoimRoleType()).thenReturn(MoimRoleType.NORMAL);
-        when(linker3.getMemberState()).thenReturn(MoimMemberState.ACTIVE);
-        sampleList.add(linker1);
-        sampleList.add(linker2);
-        sampleList.add(linker3);
-
-        // given - 사용 repository 함수 지정
-        when(moimRepository.findWithRulesById(curMoim.getId())).thenReturn(curMoim);
-        when(memberMoimLinkerRepository.findByMemberId(requestMember.getId())).thenReturn(sampleList);
-        when(memberRepository.findMemberAndMemberInfoById(requestMember.getId())).thenReturn(requestMember);
-
-        // given (request 정보)
-        MoimJoinRequestDto requestDto = new MoimJoinRequestDto(curMoim.getId());
-
-        // when
-        MyMoimLinkerDto myMoimLinkerDto = moimMemberService.requestJoin(requestDto, requestMember);
-
-        // then
-        assertEquals(MoimRoleType.NORMAL, myMoimLinkerDto.getMoimRoleType());
-        assertEquals(MoimMemberState.WAIT_BY_MOIM_CNT, myMoimLinkerDto.getMemberState());
-        assertNotNull(myMoimLinkerDto.getCreatedAt());
-    }
-
-    @Test
-    @DisplayName("실패 @ requestJoin() - 겸직으로 가입 요청 대기")
-    void 겸직_가입요청() {
-
-        // given (Rule Join 이 존재하도록 설정)
-        lenient().when(curMoimRuleJoin.isDupLeaderAvailable()).thenReturn(false);
-        lenient().when(curMoimRuleJoin.isDupManagerAvailable()).thenReturn(false);
-
-        List<MemberMoimLinker> sampleList = new ArrayList<>();
-        MemberMoimLinker linker1 = mock(MemberMoimLinker.class);
-        MemberMoimLinker linker2 = mock(MemberMoimLinker.class);
-        when(linker1.getMoimRoleType()).thenReturn(MoimRoleType.NORMAL);
-        when(linker1.getMemberState()).thenReturn(MoimMemberState.ACTIVE);
-        when(linker2.getMoimRoleType()).thenReturn(MoimRoleType.LEADER); // 리더인 모임이 있음
-        when(linker2.getMemberState()).thenReturn(MoimMemberState.ACTIVE);
-        sampleList.add(linker1);
-        sampleList.add(linker2);
-
-        // given - 사용 repository 함수 지정
-        when(moimRepository.findWithRulesById(curMoim.getId())).thenReturn(curMoim);
-        when(memberMoimLinkerRepository.findByMemberId(requestMember.getId())).thenReturn(sampleList);
-        when(memberRepository.findMemberAndMemberInfoById(requestMember.getId())).thenReturn(requestMember);
-
-        // given (request 정보)
-        MoimJoinRequestDto requestDto = new MoimJoinRequestDto(curMoim.getId());
-
-        // when
-        MyMoimLinkerDto myMoimLinkerDto = moimMemberService.requestJoin(requestDto, requestMember);
-
-        // then
-        assertEquals(MoimRoleType.NORMAL, myMoimLinkerDto.getMoimRoleType());
-        assertEquals(MoimMemberState.WAIT_BY_DUP, myMoimLinkerDto.getMemberState());
-        assertNotNull(myMoimLinkerDto.getCreatedAt());
-    }
-
-    @Test
-    @DisplayName("성공 @ requestJoin() - 재가입 시도")
-    void 재가입_가입요청() {
-
-
-    }
-
-    @Test
-    @DisplayName("실패 @ requestJoin() - 재가입 금지로 인한 실패")
-    void 재가입_금지_가입요청() {
-
-    }
 
     @Test
     @DisplayName("성공 @ decideJoin() - ")
